add useKitchenRefresh
