+++ app/type/typedetail.ts
@@ -1,0 +1,5 @@
+export type TypeDetail = {
+  order_id: number
+  product_id: number
+  quantity: number
+}
