+++ app/type/typeorder.ts
@@ -1,0 +1,5 @@
+export type TypeOrder = {
+    order_id: number
+    table_number: number
+    status: string
+  }
